move files to correct directories
