+++ test/test_utils/extract/test_totesys_processor.py
@@ -1,5 +1,5 @@
 import pytest
-from totesys_processor import ToteSysProcessor
+from src.extract.extract_utils.totesys_processor import ToteSysProcessor
 
 @pytest.fixture
 def sample_data():
